+++ app/controllers/Projects.java
@@ -2,21 +2,17 @@ package controllers;
 
 import controllers.securesocial.SecureSocial;
 import models.*;
-import play.Play;
-import play.data.validation.Required;
 import play.db.jpa.JPA;
-import play.i18n.Messages;
+import play.mvc.Before;
 import play.mvc.Controller;
-import play.mvc.Router;
+import play.mvc.Scope;
 import play.mvc.With;
 import securesocial.provider.SocialUser;
-import services.MailService;
+import securesocial.provider.UserId;
 import services.SecureUserService;
 import utils.Utils;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /**
  * Created by IntelliJ IDEA.
@@ -28,12 +24,38 @@ import java.util.Map;
 @With( SecureSocial.class )
 public class Projects extends Controller {
 
+    /**
+     * this checks whether a user has admin role to a project
+     *
+     * @throws Throwable
+     */
+    @Before(unless={"add"})
+    static void checkAccess() throws Throwable {
+        User user = User.loadBySocialUser(SecureSocial.getCurrentUser());
+        String id = Scope.Params.current().get("projectId");
+        Project project = Project.findById(Long.valueOf(id));
+        if (Participation.find("project=? and user=? and role=?", project, user, Role.ADMIN).first() == null) {
+            throw new RuntimeException("");
+        }
+    }
+
+    /**
+     * prepares for the edit setting screen
+     * @param id
+     */
     public static void edit(Long id) {
         Project project = Project.findById(id);
         List<Participation> participations = Participation.find("byProject", project).fetch();
         render(project, participations);
     }
-    
+
+    /**
+     * save the edited settings
+     *
+     * @param id
+     * @param title
+     * @param visibility
+     */
     public static void save(Long id, String title, Visibility visibility) {
         Project project = Project.findById(id);
         project.title = title;
@@ -46,7 +68,7 @@ public class Projects extends Controller {
         User user = SecureUserService.getCurrentUser();
         addProject(user, title);
         // return the whole list to refresh UI
-        renderText(Utils.toJson(getProjects(user)));
+        renderText(Utils.toJson(getParticipations(user)));
     }
 
     /**
@@ -69,26 +91,25 @@ public class Projects extends Controller {
         }
     }
 
-    public static void deleteAdmin(Long participationId) {
+    public static void deleteAdmin(Long participationId, Long projectId) {
         Participation participation = Participation.findById(participationId);
+
         // TODO check permission
-        if (participation != null && participation.role.equals(Role.ADMIN)) {
+        if (participation != null && participation.role.equals(Role.ADMIN) && 
+                participation.project.id == projectId) {
             participation.role = Role.USER;
             participation.save();
             renderText(participationId);
-        } else {
-            renderText("");
         }
     }
 
-    public static void deleteMember(Long participationId) {
+    public static void deleteMember(Long participationId, Long projectId) {
         Participation participation = Participation.findById(participationId);
         // TODO check permission
-        if (participation != null) {
+        if (participation != null &&
+                participation.project.id == projectId) {
             participation.delete();
             renderText(participationId);
-        } else {
-            renderText("");
         }
     }
 
@@ -109,11 +130,11 @@ public class Projects extends Controller {
      * @param user
      * @return
      */
-    static List<Project> getProjects(User user) {
-        List<Project> projects = JPA.em()
-                .createQuery("select proj from Participation p left join p.project proj where p.user=:user order by proj.dateCreated desc")
+    static List<Participation> getParticipations(User user) {
+        List<Participation> participations = JPA.em()
+                .createQuery("select p from Participation p left join p.project proj where p.user=:user order by proj.dateCreated desc")
                 .setParameter("user", user).getResultList();
-        return projects;
+        return participations;
     }
 
     /**
@@ -123,7 +144,7 @@ public class Projects extends Controller {
      * @param title
      * @return
      */
-    static Project addProject(User user, String title) {
+    static Participation addProject(User user, String title) {
 
         // create default project if no project is available
         Project project = new Project();
@@ -135,6 +156,6 @@ public class Projects extends Controller {
         participation.user = user;
         participation.role = Role.ADMIN;
         participation.save();
-        return project;
+        return participation;
     }
 }
